+++ .gitignore
@@ -43,7 +43,7 @@ yarn-error.log
 *.xccheckout
 *.xcscmblueprint
 xcuserdata
-packages/expo-modules-core/expo-module-optimized-macros-plugin/.build
+packages/expo-module-optimized-macros-plugin/plugin/.build
 
 # IDEA / Android Studio
 *.iml
