--- packages/expo-modules-core/expo-module-optimized-macros-plugin/Package.swift
@@ -1,43 +1,0 @@
-// swift-tools-version: 6.2
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
-import PackageDescription
-import CompilerPluginSupport
-
-let package = Package(
-    name: "ExpoModulesOptimized",
-    platforms: [.macOS(.v10_15), .iOS(.v13), .tvOS(.v13), .watchOS(.v6), .macCatalyst(.v13)],
-    products: [
-        .executable(
-            name: "ExpoModulesOptimized",
-            targets: ["ExpoModulesOptimized"]
-        ),
-    ],
-    dependencies: [
-        .package(url: "https://github.com/swiftlang/swift-syntax.git", from: "602.0.0-latest"),
-    ],
-    targets: [
-        // Targets are the basic building blocks of a package, defining a module or a test suite.
-        // Targets can depend on other targets in this package and products from dependencies.
-        // Macro implementation that performs the source transformation of a macro.
-        .macro(
-            name: "ExpoModulesOptimizedMacros",
-            dependencies: [
-                .product(name: "SwiftSyntaxMacros", package: "swift-syntax"),
-                .product(name: "SwiftCompilerPlugin", package: "swift-syntax")
-            ]
-        ),
-
-        // A client of the library, which is able to use the macro in its own code.
-        .executableTarget(name: "ExpoModulesOptimized", dependencies: ["ExpoModulesOptimizedMacros"]),
-
-        // A test target used to develop the macro implementation.
-        .testTarget(
-            name: "ExpoModulesOptimizedTests",
-            dependencies: [
-                "ExpoModulesOptimizedMacros",
-                .product(name: "SwiftSyntaxMacrosTestSupport", package: "swift-syntax"),
-            ]
-        ),
-    ]
-)
