+++ packages/expo-modules-core/package.json
@@ -55,6 +55,7 @@
     "preset": "expo-module-scripts"
   },
   "dependencies": {
+    "@expo/expo-module-optimized-macros-plugin": "~0.0.5",
     "invariant": "^2.2.4"
   },
   "peerDependencies": {
