+++ pnpm-workspace.yaml
@@ -39,3 +39,6 @@ patchedDependencies:
 preferWorkspacePackages: true
 
 minimumReleaseAge: 1440
+
+minimumReleaseAgeExclude:
+  - "@expo/*"
